Fix slightly trickier rrule() doctest
The test is basically aanother way of writing the previous example.
So for brevity it didn't include the output, but doctest struggles to tell
the difference between an ellipsis representing all the output, and ellipsis
used as a REPL line continuation.
So I've added the first and last values of the output as a visual aid to both
readers of the documentation, and doctest.